idleStateIdとwatchを移動
